+++ v2/models/loans.go
@@ -7,12 +7,12 @@ type Loans struct {
 }
 
 type Loan struct {
-	ID           int    `json:"id"`
+	ID           uint64 `json:"id"`
 	Quantity     string `json:"quantity"`
 	Rate         string `json:"rate"`
-	CreatedAt    int    `json:"created_at"`
-	LenderID     int    `json:"lender_id"`
-	BorrowerID   int    `json:"borrower_id"`
+	CreatedAt    uint   `json:"created_at"`
+	LenderID     uint   `json:"lender_id"`
+	BorrowerID   uint   `json:"borrower_id"`
 	Status       string `json:"status"`
 	Currency     string `json:"currency"`
 	FundReloaned bool   `json:"fund_reloaned"`
